style: 清理编译 warnings
- DEBUG 关闭时未使用的调试变量统一加 [[maybe_unused]]
  (axis.cpp / stepaxis.cpp / MotorControl.cpp)
- TMC2240.h R2240_70 0xC44C001E 显式 (int32_t) cast，消除 narrowing
- objectives.cpp empty-body if 加大括号；ObjectivePosition 加 [[maybe_unused]]

+++ firmware/octoaxes/tmc/ic/TMC2240/TMC2240.h
@@ -165,7 +165,7 @@ typedef struct {
 #define R2240_10 0x00070A03  // IHOLD_IRUN
 #define R2240_11 0x0000000A  // TPOWERDOWN
 #define R2240_6C 0x14410153  // CHOPCONF
-#define R2240_70 0xC44C001E  // PWMCONF
+#define R2240_70 ((int32_t)0xC44C001E)  // PWMCONF
 
 // Register access permissions
 static const uint8_t tmc2240_registerAccess[TMC2240_REGISTER_COUNT] =
